requirements.txt
pytest>=7.0
pytest-asyncio>=0.21
homeassistant
pyyaml>=6.0
numpy>=1.24
